Fix the location of the clear button.
The clearButtonRectForBounds method was returning a rect that made the
clear button appear too low in the UITextField.  Instead of adding the
entire line height to the y coordinate, it should only be adding half
of the line height so that the clear button is vertically centered with
the text.